Remove the platform implementations
