+++ ci/licenses_golden/licenses_flutter
@@ -1760,9 +1760,6 @@ FILE: ../../../flutter/shell/platform/windows/client_wrapper/include/flutter/plu
 FILE: ../../../flutter/shell/platform/windows/client_wrapper/plugin_registrar_windows_unittests.cc
 FILE: ../../../flutter/shell/platform/windows/cursor_handler.cc
 FILE: ../../../flutter/shell/platform/windows/cursor_handler.h
-FILE: ../../../flutter/shell/platform/windows/direct_manipulation.cc
-FILE: ../../../flutter/shell/platform/windows/direct_manipulation.h
-FILE: ../../../flutter/shell/platform/windows/direct_manipulation_unittests.cc
 FILE: ../../../flutter/shell/platform/windows/display_helper_winuwp.cc
 FILE: ../../../flutter/shell/platform/windows/display_helper_winuwp.h
 FILE: ../../../flutter/shell/platform/windows/dpi_utils_win32.cc
