+++ shell/platform/windows/window_binding_handler.h
@@ -29,12 +29,6 @@ struct PhysicalWindowBounds {
   size_t height;
 };
 
-// Structure containing the position of a mouse pointer
-struct PointerLocation {
-  size_t x;
-  size_t y;
-};
-
 // Type representing an underlying platform window.
 #ifdef WINUWP
 using PlatformWindow =
@@ -101,9 +95,6 @@ class WindowBindingHandler {
   // Invoked when the app ends IME composing, such when the active text input
   // client is cleared.
   virtual void OnResetImeComposing() = 0;
-
-  // Returns the last known position of the primary pointer
-  virtual PointerLocation GetPrimaryPointerLocation() = 0;
 };
 
 }  // namespace flutter
