+++ shell/platform/windows/window_binding_handler_delegate.h
@@ -50,14 +50,6 @@ class WindowBindingHandlerDelegate {
   virtual void OnPointerLeave(FlutterPointerDeviceKind device_kind,
                               int32_t device_id) = 0;
 
-  virtual void OnPointerPanZoomStart(int32_t device_id) = 0;
-  virtual void OnPointerPanZoomUpdate(int32_t device_id,
-                                      double pan_x,
-                                      double pan_y,
-                                      double scale,
-                                      double rotation) = 0;
-  virtual void OnPointerPanZoomEnd(int32_t device_id) = 0;
-
   // Notifies delegate that backing window has received text.
   // Typically called by currently configured WindowBindingHandler
   virtual void OnText(const std::u16string&) = 0;
